Add a PalettePointer to dynamically load and unload palettes in the game

+++ source/gameplay/Player.hpp
@@ -18,7 +18,7 @@ class Player final
     void listenToCommands();
     GameScene& gameScene();
 
-    u32 tileId, paletteId;
+    u16 tileId, paletteId;
 
 public:
     Player() {}
